Shift to using modified constructs, slight refactoring
